Removed winrates.png
(I'm using this as an excuse to practice full length commit messages)
In the past commits I've included winrates.png because I neglected to
delete it. There is no reason for it to be in the repository, it is
simply created as a result of testing.

The problem was that I've got an uneccesarily complicated directory here
and I attempted to fix it in a simple way by removing an uneccesary file.