Fix match volume increase
Previous commit was a very broken version, fixed several problems:

getMatchHistory now takes an optional parameter, an int match id.
It uses the same API call but with a less_than_match_id param.
This fixed a problem caused by the fact that the publicMatches call
does not return new matches for some time.

Also added a couple of lines after the win numbers are printed to
convert the numbers into percentages so they can be graphed.