Begin project
Added several files with flexible names and contents.

Began initial work on computer.rb. Currently working towards defining the
computer type independently of the scraper.

In the future, I will be implementing one or more scrapers to pull information
from the web and a main class for the operation of the tool and the creation
of the storage file.