move serialization to marshal
using marshal for serialization now, much better in every way.

also created directory test for file test.rb which creates a sample computer
and tests the methods on it.

currently there is a bug with the way i am storing the object, not sure
which file it's in but it will be an easy fix (name is the processor hash
for some reason)