fix create and load bugs
the create bug was caused by me forgetting to account for the name of the
computer within create_computer. resolved by adding another parameter for
names. may eventually decide to just name the name part of the details hash
that goes into the create_computer function.

the load bug was caused by the existence of the old computer within the file
and the fact that load_computer actually loads the entire file. Deleting the
file resolved the issue partially, but I will have to make changes to
load_computer so that it actually picks out just the one computer.