Implement commmand line entry
made changes to test.rb to allow command line entry of computers. In the
future, I would like to make the entry system much more robust and useful.

Added a function to computers.rb that allows a user to delete all computers
by deleting computers.cstore. This function has three-factor authentication
but is NOT secure in any way because the user can read the password in the
file. The user can also just delete the file manually to accomplish the same
goal.